pyproject.toml
[tool.poetry]
name = "Transfer"
version = "0.1.0"
description = ""
authors = ["André Felipe Dias <andref.dias@gmail.com>"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
aiofiles = "*"
apscheduler = "*"
fastapi = "*"
hypercorn = "*"
loguru = "*"
orjson = "*"
redis = { extras = ["hiredis"], version = "*" }
python = "^3.11"
python-dotenv = "*"
python-multipart = "*"
tenacity = "*"
uvloop = "*"

[tool.poetry.group.dev.dependencies]
asgi-lifespan = "*"
alt-pytest-asyncio = "*"
blue = "*"
httpx = "*"
mypy = "*"
pip-audit = "*"
pytest = "*"
pytest-cov = "*"
ruff = "*"
types-aiofiles = "*"
types-redis = "*"

[tool.blue]
line-length = 100

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:The localize method is no longer necessary",
]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
select = [
    "ARG",
    "B",
    "C4",
    "E",
    "ERA",
    "F",
    "I",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "RET",
    "S",
    "T10",
    "T10",
    "UP",
    "W",
]
ignore = ["B008"]
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["ARG", "PLR2004", "S"]
"smoke_test.py" = ["S"]
